requirements.txt
streamlit
openai
audio_recorder_streamlit
python-dotenv
streamlit-float